feat: Support actual paths (#230)
## What?

This PR allows actual `syn::Path`s to be used as the crate path. It also
is backward compatible with the current strings-as-paths system.

## Why?

`macro_rules` macros create special tokens which cannot be recreated
from a string. To support them in paths, namely, the path to
`externref`, we would need to support actual paths instead of path
strings.